@@ -1,42 +1,48 @@
 name: awareframework_activityrecognition
-description: An Aware Framework Plugin for Monitoring Motion Activities
-version: 0.0.5
-author: AWARE Mobile Context Instrumentation Middleware Framework <yuuki.nishiyama@oulu.fi>
+description: A new flutter plugin project.
+version: 0.1.0
 homepage: http://www.awareframework.com
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  awareframework_core: ^0.1.1
-
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
+  awareframework_core: ^0.6.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test:
+
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
+  # This section identifies this Flutter project as a plugin project.
+  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
+  # be modified. They are used by the tooling to maintain consistency when
+  # adding or updating assets for this project.
   plugin:
-    androidPackage: com.awareframework.activityrecognition.awareframeworkactivityrecognition
-    pluginClass: AwareframeworkActivityrecognitionPlugin
+    platforms:
+      android:
+        package: com.awareframework.activityrecognition.awareframework_activityrecognition
+        pluginClass: AwareframeworkActivityrecognitionPlugin
+      ios:
+        pluginClass: AwareframeworkActivityrecognitionPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
   #
   # For details regarding assets in packages, see
-  # https://flutter.io/assets-and-images/#from-packages
+  # https://flutter.dev/assets-and-images/#from-packages
   #
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.io/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware.
 
   # To add custom fonts to your plugin package, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
@@ -56,4 +62,4 @@ flutter:
   #         weight: 700
   #
   # For details regarding fonts in packages, see
-  # https://flutter.io/custom-fonts/#from-packages
+  # https://flutter.dev/custom-fonts/#from-packages
